+++ app/public/index.html
@@ -6,10 +6,16 @@
     <meta name="theme-color" content="#000000" />
     <meta name="description" content="Chart Challenge" />
     <title>Chart Challenge</title>
-    <!-- <link
+    <link rel="preconnect" href="https://fonts.gstatic.com" />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Source+Sans+Pro&display=swap"
       rel="stylesheet"
-      href="//cdnjs.cloudflare.com/ajax/libs/highlight.js/10.7.2/styles/default.min.css"
-    /> -->
+    />
+    <link rel="preconnect" href="https://fonts.gstatic.com" />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Source+Code+Pro&family=Source+Sans+Pro&display=swap"
+      rel="stylesheet"
+    />
   </head>
   <body class="d-flex flex-column h-100">
     <noscript>You need to enable JavaScript to run this app.</noscript>
