--- app/src/app.scss
@@ -1,34 +1,0 @@
-header {
-  background: #dddee1;
-
-  & .container {
-    height: 65px;
-  }
-}
-
-footer {
-  background: #dddee1;
-  position: fixed;
-  left: 0;
-  right: 0;
-  bottom: 0;
-
-  & .container {
-    height: 65px;
-  }
-}
-
-.editor {
-  //   border-radius: 6px;
-  //   box-shadow: 0 2px 2px 0 rgba(0, 0, 0, 0.14), 0 1px 5px 0 rgba(0, 0, 0, 0.12),
-  //     0 3px 1px -2px rgba(0, 0, 0, 0.2);
-  //   font-family: "Source Code Pro", monospace;
-  //   font-size: 14px;
-  //   font-weight: 400;
-  //   height: 340px;
-  //   letter-spacing: normal;
-  //   line-height: 20px;
-  //   padding: 10px;
-  //   tab-size: 4;
-  //   background: #000 !important;
-}
