+++ app/src/chart.scss
@@ -1,0 +1,60 @@
+* {
+  font-family: "Source Sans Pro", sans-serif;
+}
+
+body {
+  padding: 0;
+  width: 100%;
+}
+
+#root {
+  width: 100%;
+}
+
+header {
+  background: #dddee1;
+  font-size: 35px;
+  height: 65px;
+  display: flex;
+  flex-flow: column;
+  justify-content: center;
+  align-items: flex-start;
+}
+
+main {
+  height: 100%;
+  max-height: 500px;
+}
+
+footer {
+  background: #dddee1;
+  position: fixed;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  height: 65px;
+
+  display: flex;
+  flex-flow: column;
+  justify-content: center;
+  align-items: flex-start;
+
+  & .btn {
+    width: 120px;
+    padding: 8px;
+  }
+}
+
+.Codemirror {
+  border-radius: 6px;
+  box-shadow: 0 2px 2px 0 rgba(0, 0, 0, 0.14), 0 1px 5px 0 rgba(0, 0, 0, 0.12),
+    0 3px 1px -2px rgba(0, 0, 0, 0.2);
+  font-family: "Source Code Pro", monospace;
+  font-size: 14px;
+  font-weight: 400;
+  height: 340px;
+  letter-spacing: normal;
+  line-height: 20px;
+  padding: 10px;
+  tab-size: 4;
+}
