+++ app/src/index.js
@@ -1,17 +1,14 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import "./app.scss";
+import { Context } from "./context";
+
 import App from "./App";
-import reportWebVitals from "./reportWebVitals";
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <Context.Provider value={Context}>
+      <App />
+    </Context.Provider>
   </React.StrictMode>,
   document.getElementById("root")
 );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
